llamaindex/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
llama-index
llama-index-embeddings-google-genai
llama-index-llms-openai
llama-index-core
sqlalchemy
python-dotenv
